# state: Copy data on write - not on read