Depends on https://github.com/hashicorp/terraform-schema/pull/39

Closes #471

--- 

## Benchmark Results

### Before (current `main`)

```
go test -run=XXX -bench=. ./internal/state
```
```
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/terraform-ls/internal/state
cpu: Intel(R) Core(TM) i7-7567U CPU @ 3.50GHz
BenchmarkModuleByPath-4     	    3937	    268528 ns/op
BenchmarkProviderSchema-4   	   58588	     20559 ns/op
PASS
ok  	github.com/hashicorp/terraform-ls/internal/state	3.495s
```

### After (branch linked to this PR)

```
go test -run=XXX -bench=. ./internal/state
```
```
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/terraform-ls/internal/state
cpu: Intel(R) Core(TM) i7-7567U CPU @ 3.50GHz
BenchmarkModuleByPath-4     	    9272	    116078 ns/op
BenchmarkProviderSchema-4   	  104325	     11234 ns/op
PASS
ok  	github.com/hashicorp/terraform-ls/internal/state	2.537s
```